third_party/blink/web_tests/external/wpt/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-source-media-env-change-expected.txt

This is a testharness.js-based test.
FAIL the <source> media attribute: no reaction to environment change assert_equals: afterLoadCalled expected "#b" but got "#a"
Harness: the test ran to completion.
